refactor(library/ShimmedStabilityAIClient): aligns file structure with namespaces in "ShimmedStabilityAIClient/toShortfin"
