+++ src/library/ShimmedStabilityAIClient/toShortfinRequestBody/Prompt.ts
