+++ src/library/ShimmedStabilityAIClient/toShortfinRequestBody/index.ts
